requirements.txt
pyyaml

requests
fastapi
fastapi-utils[all]
influxdb-client

smbus2
gpiozero
ADS1x15-ADC
adafruit-circuitpython-ms8607
adafruit-circuitpython-as726x
adafruit-circuitpython-mlx90640
pimoroni-sgp30
opencv-python
psutil
